+++ front/src/app/App.tsx
@@ -1,16 +1,10 @@
 import React, { useLayoutEffect } from "react";
 import { useDispatch } from "react-redux";
-import { CrispChat, MaintenanceCallout } from "react-design-system";
-import {
-  useFeatureFlags,
-  useFetchFeatureFlags,
-} from "src/app/hooks/useFeatureFlags";
+import { CrispChat } from "react-design-system";
+import { useFetchFeatureFlags } from "src/app/hooks/useFeatureFlags";
 import { MatomoTagManager } from "src/app/MatomoTagManager";
 import { ENV } from "src/config/environmentVariables";
 import { appIsReadyAction } from "src/core-logic/domain/actions";
-import { featureFlagSelectors } from "src/core-logic/domain/featureFlags/featureFlags.selector";
-import { commonContent } from "./contents/commonContent";
-import { useAppSelector } from "./hooks/reduxHooks";
 import { Router } from "./routes/Router";
 
 const useAppIsReady = () => {
@@ -23,21 +17,8 @@ const useAppIsReady = () => {
 export const App = () => {
   useFetchFeatureFlags();
   useAppIsReady();
-  const { enableMaintenance } = useFeatureFlags();
-  const maintenanceMessage = useAppSelector(
-    featureFlagSelectors.maintenanceMessage,
-  );
   return (
     <>
-      {enableMaintenance.isActive && (
-        <MaintenanceCallout
-          message={
-            maintenanceMessage || maintenanceMessage !== ""
-              ? maintenanceMessage
-              : commonContent.maintenanceMessage
-          }
-        />
-      )}
       <Router />
       {ENV.crispWebSiteId && <CrispChat crispWebsiteId={ENV.crispWebSiteId} />}
